Completed conditional shows of events.
Learned about v-if and setting booleans to the inverse of what they currently are for toggle functionality.

This can also be done in a v-if evaluating a boolean to determine whether it shows when true or shows when false.

The difference between v-if and v-show is v-if removes and inserts the conditional elements when conditions are met. Whereas v-show simply uses style rules to change the element from its current display property to none.