Created Dynamic Routes
Created routes for each job dynamically by passing the Job ID as a prop to the JobDetails view.

A job will now show content that matches the job ID in the id prop passed to it.

Added styling to the job links. Moving to make redirects and a 404 page for the website.