Added JSON Server DB, data mounting and iteration
Added JSON Server and make Projects object that contains two sample projects.

In HomeView, retrieved the JSON output of the data and assigned it to projects data property for iteration.

V-if and V-for directives allow the data to show if the array is populated and iterate through it using a single component.